chore: ruff config
Adds [tool.ruff] with target-version py311, line-length 100, and a sensible
lint set (pycodestyle, pyflakes, isort, bugbear, pyupgrade, ruff-specific).

Switches the CLI enums to StrEnum (Python 3.11+ idiom; the str+Enum mixin
predates StrEnum) and drops a redundant int(round(...)) in the audio test.

B008 is per-file-ignored for cli.py since typer.Option() in argument defaults
is the canonical Typer pattern.

+++ tests/test_audio.py
@@ -34,7 +34,7 @@ def test_ensure_16k_resamples_22k_to_16k(sine_wav_22k: Path) -> None:
     samples, sr = load_audio(sine_wav_22k)
     assert sr == 22_050
     out = ensure_16k(samples, sr)
-    expected_len = int(round(len(samples) * TARGET_SR / sr))
+    expected_len = round(len(samples) * TARGET_SR / sr)
     # torchaudio's resampler may differ by a few samples vs the naive ratio.
     assert abs(len(out) - expected_len) <= 4
     assert out.dtype == np.float32
